Add results command
This command receives a job name and an optional build number. It
returns the resulting cases for the specified job and build. If the
build number is omitted, it returns the corresponding results for the
last build of the specified job. If the build is still running, it
raises an error.